+++ lib/src/math/algebra/expression/components/functions/trig/cot.dart
@@ -33,8 +33,25 @@ class Cot extends TrigonometricExpression {
 
   @override
   Expression integrate() {
-    // Integral of cot(x) is ln|sin(x)|
-    return Ln(Abs(Sin(operand)));
+    // ∫cot(ax + b) dx = (1/a) * ln|sin(ax + b)|
+    dynamic a = 1;
+
+    if (operand is Add) {
+      if ((operand as Add).left is Multiply) {
+        a = ((operand as Add).left as Multiply).left.evaluate();
+      }
+    } else if (operand is Multiply) {
+      a = (operand as Multiply).left.evaluate();
+    }
+
+    dynamic invA;
+    if (a is Complex) {
+      invA = Complex.one() / a;
+    } else {
+      invA = 1 / a;
+    }
+
+    return Multiply(Literal(invA), Ln(Abs(Sin(operand))));
   }
 
   @override