+++ lib/src/math/algebra/expression/components/functions/trig/sin.dart
@@ -23,7 +23,25 @@ class Sin extends TrigonometricExpression {
 
   @override
   Expression integrate() {
-    return Negate(Cos(operand));
+    // ∫sin(ax + b) dx = -(1/a) * cos(ax + b)
+    dynamic a = 1;
+
+    if (operand is Add) {
+      if ((operand as Add).left is Multiply) {
+        a = ((operand as Add).left as Multiply).left.evaluate();
+      }
+    } else if (operand is Multiply) {
+      a = (operand as Multiply).left.evaluate();
+    }
+
+    dynamic negInvA;
+    if (a is Complex) {
+      negInvA = Complex(-1, 0) / a;
+    } else {
+      negInvA = -1 / a;
+    }
+
+    return Multiply(Literal(negInvA), Cos(operand));
   }
 
   @override