+++ lib/src/math/algebra/expression/components/functions/trig/tan.dart
@@ -18,7 +18,25 @@ class Tan extends TrigonometricExpression {
 
   @override
   Expression integrate() {
-    return Negate(Ln(Abs(Cos(operand))));
+    // ∫tan(ax + b) dx = -(1/a) * ln|cos(ax + b)|
+    dynamic a = 1;
+
+    if (operand is Add) {
+      if ((operand as Add).left is Multiply) {
+        a = ((operand as Add).left as Multiply).left.evaluate();
+      }
+    } else if (operand is Multiply) {
+      a = (operand as Multiply).left.evaluate();
+    }
+
+    dynamic negInvA;
+    if (a is Complex) {
+      negInvA = Complex(-1, 0) / a;
+    } else {
+      negInvA = -1 / a;
+    }
+
+    return Multiply(Literal(negInvA), Ln(Abs(Cos(operand))));
   }
 
   @override